refactor: extract Field to field package
