+++ packages/_shared/src/index.tsx
@@ -21,7 +21,6 @@ export { FieldConnector } from './FieldConnector';
 export { PredefinedValuesError } from './PredefinedValuesError';
 export { CharCounter } from './CharCounter';
 export { CharValidation } from './CharValidation';
-export { Field } from './Field';
 
 import * as ModalDialogLauncher from './ModalDialogLauncher';
 export { ModalDialogLauncher };
