+++ packages/field/.eslintrc.js
@@ -1,0 +1,8 @@
+const path = require('path');
+
+module.exports = {
+  extends: '../.eslintrc.js',
+  parserOptions: {
+    project: path.resolve('tsconfig.json'),
+  },
+};
