+++ packages/field/README.md
@@ -1,0 +1,11 @@
+# @contentful/field-editor-field
+
+```bash
+npm install @contentful/field-editor-field
+```
+
+This package contains a React `Field` component which is used to render different field editors based on the widget ID.
+
+```js
+import { Field } from '@contentful/field-editor-field';
+```
