+++ packages/field/package.json
@@ -1,0 +1,61 @@
+{
+  "name": "@contentful/field-editor-field",
+  "version": "0.1.1",
+  "main": "dist/index.js",
+  "module": "dist/field-editor-field.esm.js",
+  "typings": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "author": "Contentful GmbH",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/contentful/field-editors"
+  },
+  "scripts": {
+    "watch": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test --env=jsdom --watch",
+    "test:ci": "tsdx test --env=jsdom --ci",
+    "tsc": "tsc -p ./ --noEmit"
+  },
+  "dependencies": {
+    "contentful-ui-extensions-sdk": "^3.18.0",
+    "@contentful/field-editor-number": "^0.9.0",
+    "@contentful/field-editor-single-line": "^0.10.0",
+    "@contentful/field-editor-boolean": "^0.8.0",
+    "@contentful/field-editor-date": "^0.7.0",
+    "@contentful/field-editor-location": "^0.8.0",
+    "@contentful/field-editor-json": "^0.7.0",
+    "@contentful/field-editor-multiple-line": "^0.9.0",
+    "@contentful/field-editor-tags": "^0.9.0",
+    "@contentful/field-editor-slug": "^0.6.0",
+    "@contentful/field-editor-dropdown": "^0.10.0",
+    "@contentful/field-editor-url": "^0.8.0",
+    "@contentful/field-editor-radio": "^0.9.0",
+    "@contentful/field-editor-rating": "^0.8.0",
+    "@contentful/field-editor-checkbox": "^0.10.1",
+    "@contentful/field-editor-list": "^0.11.0",
+    "@contentful/field-editor-reference": "^1.1.2",
+    "@contentful/field-editor-rich-text": "^0.14.2",
+    "@contentful/field-editor-markdown": "^0.9.0",
+    "@contentful/field-editor-shared": "^0.12.0"
+  },
+  "devDependencies": {
+    "@contentful/field-editor-test-utils": "^0.8.0"
+  },
+  "peerDependencies": {
+    "react": ">=16.8.0"
+  },
+  "jest": {
+    "testMatch": [
+      "**/?(*.)+(spec|test).[jt]s?(x)"
+    ],
+    "globals": {
+      "ts-jest": {
+        "diagnostics": false
+      }
+    }
+  }
+}
